feat: workspace and different stages as packages
