+++ fundamentals/src/asset.rs
