+++ fundamentals/src/constants.rs
@@ -1,0 +1,1 @@
+pub const ALICE: sp_runtime::AccountId32 = sp_runtime::AccountId32::new([1u8; 32]);
