+++ fundamentals/src/lib.rs
@@ -1,0 +1,31 @@
+// Lesson 1
+#[cfg(feature = "location")]
+pub mod location;
+
+// Lesson 2
+#[cfg(feature = "asset")]
+pub mod asset;
+
+// Lesson 3
+#[cfg(feature = "instruction")]
+pub mod instruction;
+
+// Lesson 4
+#[cfg(feature = "holding")]
+pub mod holding;
+
+// Lesson 5
+#[cfg(feature = "xcm_executor")]
+pub mod xcm_executor;
+
+// Lesson 6
+#[cfg(feature = "pallet_xcm")]
+pub mod pallet_xcm;
+
+#[cfg(feature = "pallet_xcm")]
+pub mod network;
+
+mod constants;
+
+#[cfg(test)]
+mod tests;
