--- src/parachain/xcm_config/reserve.rs
@@ -1,15 +1,0 @@
-pub use sandbox::*;
-
-#[cfg(feature = "start")]
-mod sandbox {
-	pub type TrustedReserves = ();
-}
-
-#[cfg(feature = "example")]
-mod sandbox {
-	use crate::parachain::teleporter::TrustedTeleporters;
-	use frame_support::traits::EverythingBut;
-	use xcm_builder::NativeAsset;
-
-	pub type TrustedReserves = (NativeAsset, EverythingBut<TrustedTeleporters>);
-}
