--- src/tests/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "example")]
-mod example;
-
-#[cfg(feature = "fundamentals")]
-mod fundamentals;
